Revert "updating imports and moving hydrogen tank to unit_models folder"
This reverts commit 76a7dc96a46da93ef78ea79743de5a1cfef0c684.
