+++ dispatches/models/nuclear_case/test_hydrogen_tank.py
@@ -37,10 +37,8 @@ from idaes.generic_models.properties.core.generic.generic_property \
 from idaes.core.util.model_statistics import degrees_of_freedom
 from idaes.core.util.testing import initialization_tester
 # Import unit model and property package
-from dispatches.models.nuclear_case.unit_models.hydrogen_tank \
-    import HydrogenTank
-from dispatches.models.nuclear_case.properties.hturbine_ideal_vap \
-    import configuration
+from hydrogen_tank import HydrogenTank
+from h2_ideal_vap import configuration
 
 # Get default solver for testing
 solver = get_solver()
@@ -50,14 +48,14 @@ def test_config():
     # Create the ConcreteModel and the FlowSheetBlock
     m = ConcreteModel(name="H2TankModel")
     m.fs = FlowsheetBlock(default={"dynamic": False})
-
+    
     # Add hydrogen property parameter block
     m.fs.properties = GenericParameterBlock(default=configuration)
-
+    
     # Create an instance of the CHG tank
     m.fs.unit = HydrogenTank(default={"property_package": m.fs.properties,
                                       "dynamic": False})
-
+    
     # Check unit config arguments
     assert len(m.fs.unit.config) == 5
 
@@ -73,10 +71,10 @@ class TestH2IdealVap(object):
         # Create the ConcreteModel and the FlowSheetBlock
         m = ConcreteModel(name="H2TankModel")
         m.fs = FlowsheetBlock(default={"dynamic": False})
-
+        
         # Add hydrogen property parameter block
         m.fs.properties = GenericParameterBlock(default=configuration)
-
+        
         # Create an instance of the Hydrogen tank
         m.fs.unit = HydrogenTank(default={"property_package": m.fs.properties,
                                           "dynamic": False})
@@ -84,27 +82,27 @@ class TestH2IdealVap(object):
         # Fix tank geometry
         m.fs.unit.tank_diameter.fix(0.1)
         m.fs.unit.tank_length.fix(0.3)
-
+        
         # Fix initial state of tank
         m.fs.unit.previous_state[0].temperature.fix(300)
         m.fs.unit.previous_state[0].pressure.fix(1e5)
-
+        
         # Fix inlet state
         m.fs.unit.control_volume.properties_in[0].flow_mol.fix(100)
         m.fs.unit.control_volume.properties_in[0].mole_frac_comp.fix(1)
         m.fs.unit.control_volume.properties_in[0].temperature.fix(300)
         m.fs.unit.control_volume.properties_in[0].pressure.fix(1.10325e5)
-
+        
         # Fix Duration of Operation (Time Step)
         m.fs.unit.dt[0].fix(100)
-
+        
         # Fix the outlet flow to zero for tank filling type operation
         m.fs.unit.control_volume.properties_out[0].flow_mol.fix(0)
-
+        
         # Setting the bounds on the state variables
         m.fs.unit.control_volume.properties_in[0].pressure.setub(1e15)
         m.fs.unit.control_volume.properties_out[0].pressure.setub(1e15)
-
+        
         return m
 
     @pytest.mark.unit
@@ -121,12 +119,12 @@ class TestH2IdealVap(object):
         assert isinstance(hydrogentank.fs.unit.heat_duty, Var)
         assert isinstance(hydrogentank.fs.unit.previous_material_holdup, Var)
         assert isinstance(hydrogentank.fs.unit.previous_energy_holdup, Var)
-
+    
     # Check degrees of freedom
     @pytest.mark.unit
     def test_dof(self, hydrogentank):
         assert degrees_of_freedom(hydrogentank) == 0
-
+    
     # Check initialization
     @pytest.mark.skipif(solver is None, reason="Solver not available")
     @pytest.mark.component
@@ -154,7 +152,7 @@ class TestH2IdealVap(object):
         assert (value(hydrogentank.fs.unit.control_volume.properties_out[0].\
                       dens_mol_phase["Vap"])
                 == pytest.approx(4244171.91, rel=1e-1))
-
+    
     # Try another inlet temperature
     @pytest.mark.skipif(solver is None, reason="Solver not available")
     @pytest.mark.component
@@ -163,7 +161,7 @@ class TestH2IdealVap(object):
         results = solver.solve(hydrogentank)
         assert (results.solver.termination_condition ==
                 TerminationCondition.optimal)
-
+        
         assert (value(hydrogentank.fs.unit.control_volume.properties_out[0].pressure)
                 == pytest.approx(15536853611.796, rel=1e-1))
         assert (value(hydrogentank.fs.unit.control_volume.properties_out[0].temperature)
